Display posts as cards on the tag index page
