+++ pages/tags/[tag].tsx
@@ -1,16 +1,17 @@
 import { GetStaticPaths, GetStaticProps } from 'next'
-import Link from 'next/link'
 import React from 'react'
 import PageLayout from '../../components/PageLayout'
 import collectTags from '../../util/collectTags'
 import getQueryParam from '../../util/getQueryParam'
 import loadAllPosts from '../../util/loadAllPosts'
-import { LinkEntry } from '../posts'
 import startCase from 'lodash/startCase'
+import { MarkdownFileWithUrl } from '..'
+import omitUndefinedFields from '../../util/omitUndefinedFields'
+import ContentCardList from '../../components/ContentCardList'
 
 interface Props {
     tag: string
-    links: LinkEntry[]
+    posts: MarkdownFileWithUrl[]
 }
 
 export const getStaticPaths: GetStaticPaths = async () => {
@@ -23,14 +24,11 @@ export const getStaticProps: GetStaticProps<Props> = async context => {
     const tag = getQueryParam(context.params, 'tag').toLowerCase()
     const posts = await loadAllPosts()
     const filteredPosts = posts.filter(post => post.frontMatter.tags.includes(tag))
-    const links = filteredPosts.map(post => ({
-        title: post.frontMatter.title,
-        url: `/posts/${post.slug}`,
-    }))
+
     return {
         props: {
             tag,
-            links,
+            posts: filteredPosts.map(post => omitUndefinedFields({ ...post, url: `/posts/${post.slug}` })),
         },
     }
 }
@@ -40,15 +38,8 @@ export default function TagPage(props: Props) {
     return (
         <PageLayout contentTitle={`Posts tagged ${tagName}`}>
             <h1 className="mb-5">Posts tagged {tagName}</h1>
-            <ul>
-                {props.links.map(link => (
-                    <li>
-                        <Link href={link.url}>
-                            <a>{link.title}</a>
-                        </Link>
-                    </li>
-                ))}
-            </ul>
+
+            <ContentCardList posts={props.posts} />
         </PageLayout>
     )
 }
